+++ crawler/src/Config.h
@@ -14,6 +14,12 @@ struct CrawlerConfig {
     std::vector<std::string> seed_urls;
 
     std::string data_directory = "data";
+
+    long default_crawl_delay_ms = 200;  // milliseconds
+    size_t middle_queue_queue_count = 100;
+    size_t middle_queue_url_batch_size = 10;
+    size_t middle_queue_host_url_limit = 25;
+    double middle_queue_utilization_target = 0.25;
 };
 
 CrawlerConfig LoadConfigFromFile(const std::string& path);