+++ crawler/src/Coordinator.cpp
@@ -55,8 +55,7 @@ Coordinator::Coordinator(const CrawlerConfig& config) : config_(config) {
 
     docQueue_ = core::UniquePtr<DocumentQueue>(new DocumentQueue{state_->threadSync});
     frontier_ = core::UniquePtr<UrlFrontier>(new UrlFrontier{frontierDirectory_});
-    requestManager_ = core::UniquePtr<RequestManager>(
-        new RequestManager{config_.concurrent_requests, config_.request_timeout, frontier_.Get(), docQueue_.Get()});
+    requestManager_ = core::UniquePtr<RequestManager>(new RequestManager{frontier_.Get(), docQueue_.Get(), config});
 
     RecoverState();
 }
@@ -103,10 +102,12 @@ void Coordinator::Run() {
     sigaddset(&signals, SIGTERM);
     pthread_sigmask(SIG_BLOCK, &signals, nullptr);
 
-    int sig;
-    sigwait(&signals, &sig);
+    int sig = 0;
+    while (sig != SIGINT && sig != SIGTERM) {
+        sigwait(&signals, &sig);
+    }
 
-    spdlog::info("received signal {}, shutting down", strsignal(sig));
+    spdlog::info("received signal {} {}, shutting down", sig, strsignal(sig));
 
     // Send shutdown to threads
     state_->threadSync.Shutdown();
@@ -135,9 +136,12 @@ void Coordinator::DumpState() {
     PersistentState state;
     state.nextDocumentID = state_->nextDocumentID.load();
     frontier_->DumpPendingURLs(state.pendingURLs);
+    requestManager_->ExtractQueuedURLs(state.activeCrawlURLs);
+    docQueue_->ExtractCompletedURLs(state.activeCrawlURLs);
 
     spdlog::debug("saved state: next document id = {}", state.nextDocumentID);
     spdlog::debug("saved state: pending url count = {}", state.pendingURLs.size());
+    spdlog::debug("saved state: active crawl url count = {}", state.activeCrawlURLs.size());
 
     {
         // Write state to file
@@ -168,10 +172,11 @@ void Coordinator::RecoverState() {
 
     spdlog::debug("loaded state: next document id = {}", state.nextDocumentID);
     spdlog::debug("loaded state: pending url count = {}", state.pendingURLs.size());
+    spdlog::debug("loaded state: active crawl url count = {}", state.activeCrawlURLs.size());
 
     state_->nextDocumentID.store(state.nextDocumentID);
     frontier_->PushURLs(state.pendingURLs);
+    requestManager_->RestoreQueuedURLs(state.activeCrawlURLs);
 }
-
 
 }  // namespace mithril