+++ crawler/src/RequestManager.cpp
@@ -7,7 +7,6 @@
 #include "http/RequestExecutor.h"
 #include "http/URL.h"
 
-#include <atomic>
 #include <cassert>
 #include <cstddef>
 #include <string>
@@ -17,13 +16,10 @@
 
 namespace mithril {
 
-RequestManager::RequestManager(size_t targetConcurrentReqs,
-                               unsigned long requestTimeout,
-                               UrlFrontier* frontier,
-                               DocumentQueue* docQueue)
-    : targetConcurrentReqs_(targetConcurrentReqs),
-      requestTimeout_(requestTimeout),
-      frontier_(frontier),
+RequestManager::RequestManager(UrlFrontier* frontier, DocumentQueue* docQueue, const CrawlerConfig& config)
+    : targetConcurrentReqs_(config.concurrent_requests),
+      requestTimeout_(config.request_timeout),
+      middleQueue_(frontier, config),
       docQueue_(docQueue) {}
 
 void RequestManager::Run(ThreadSync& sync) {
@@ -40,7 +36,7 @@ void RequestManager::Run(ThreadSync& sync) {
 
             size_t toAdd = targetConcurrentReqs_ - requestExecutor_.InFlightRequests();
             urls.clear();
-            frontier_->GetURLs(sync, toAdd, urls, wantAtLeastOne);
+            middleQueue_.GetURLs(sync, toAdd, urls, wantAtLeastOne);
             if (sync.ShouldSynchronize()) {
                 continue;
             }
@@ -89,13 +85,18 @@ void RequestManager::Run(ThreadSync& sync) {
     spdlog::info("request manager terminating");
 }
 
-void RequestManager::Stop() {
-    stopped_.store(true, std::memory_order_release);
-}
-
 void RequestManager::DispatchFailedRequest(http::FailedRequest failed) {
     spdlog::warn("failed crawl request: {} {}", failed.req.Url().url, http::StringOfRequestError(failed.error));
     // TODO: pass off to whatever
+}
+
+void RequestManager::RestoreQueuedURLs(std::vector<std::string>& urls) {
+    middleQueue_.RestoreFrom(urls);
+}
+
+void RequestManager::ExtractQueuedURLs(std::vector<std::string>& out) {
+    middleQueue_.ExtractQueuedURLs(out);
+    requestExecutor_.DumpUnprocessedRequests(out);
 }
 
 }  // namespace mithril