+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - port index constructor and close the loop with crawler (#38)
 - static ranking for crawler (#41)
 - persistent, interruptible crawling (#45)
+- middle queue for rate limiting (#49)
 
 ### Fixed
 - Fix copying body data from buffer (#24)